+++ application/controllers/lists.php
@@ -47,7 +47,7 @@ class Lists extends CI_Controller
     if($subscription) { $userId = $this->session->userdata('userId'); }
     $this->data['items'] = $this->stream_model->search($search, $where, $sort, $results, $comments, $limit, $current, $userId, $subscription);
     $count = $this->stream_model->search_count($search, $where, $sort, $results, $comments, $limit, $current, $userId, $subscription);
-    $this->data['pages'] = ceil(($count->items) / $limit);
+    $this->data['pages'] = ceil(($count) / $limit);
     $this->data['title'] = "Search";
     if($search) { $this->data['title'] .= " [".$search."]"; }
     //print_r($this->data); die();