+++ CHANGELOG.md
@@ -1,13 +1,15 @@
 <h1>Changelog</h1>
 
+<h2>Version 0.9.6.2</h2>
+
+<h4>Fixes:</h4>
+* Fix count for non-search
+
 <h2>Version 0.9.6.1</h2>
 
 <h4>Fixes:</h4>
 * Keep both search fields in sync
 * Mobile Landscape Tutorial Pop-Up size
-
-<h4>Notes:</h4>
-* "Modify Comment on Mobile" bug not found
 
 <h2>Version 0.9.6.0</h2>